requirements.txt
streamlit
torch
torchvision
pillow
transformers
matplotlib
spacy
#python -m spacy download en_core_web_sm 
scikit-learn